Move code to helloworld package
This helps MyPy work correctly and allows import statements for tests to
correctly reference package and module names.
